langgraph-fin-agent/pyproject.toml
[tool.poetry]
name = "langgraph-fin-agent"
version = "0.1.0"
description = "Finance agent built using Langgraph"
authors = ["Yi Zhang <yi@relari.ai>", "Pasquale Antonante <pasquale@relari.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
langgraph = "^0.2.45"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.8"
relari-otel = "0.0.1"
beautifulsoup4 = "^4.13.3"
openinference-instrumentation-langchain = "^0.1.35"
arize-otel = "^0.8.1"
openinference-instrumentation-openai = "^0.1.29"
arize-phoenix = "^10.5.0"
arize-phoenix-otel = "^0.10.1"
langchain = "^0.3.25"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
finchat = "langgraph_fin_agent.main:main"